Pull Request 2
- Moved test functions to exportertestbase
- Fixed the test cases for the basic blenshape test.
